further API simplfiication
maxbuffers and maxrequests don't need to be passed in to each call
to dns_dispatch_createtcp() and _createudp(). in fact it's sufficient
now to use default values.

in all current uses, the value for mgr->maxbuffers will either be raised
once from its default of 20000 to 32768, or else left alone. (passing in
a value lower than 20000 does not lower it.) there isn't enough difference
between these values for there to be any need for configuration.

the value for disp->maxrequests controls both the quota of concurrent
requests for a dispatch and also the size of the dispatch socket memory
pool. it's not clear that this quota is necessary at all. the memory pool
size currently starts at 32768 and is sometimes lowered to 4096, which
is definitely unnecessary.

this commit sets both values permanently to 32768.