creating TCP dispatch now creates/binds the socket
Previously, creation of TCP dispatches differed from UDP in
that a TCP dispatch was created to attach to an existing socket,
whereas a UDP dispatch would be created in a vacuum and sockets
would be opened on demand when a transaction was initiated.

We are moving as much socket code as possible into the dispatch
module, so that it can be replaced with a netmgr version as easily
as possible. (This will also have the side effect of making TCP and
UDP dispatches more similar.)

As a step in that direction, this commit changes
dns_dispatch_createtcp() so that it creates the TCP socket. (The
still has to initiate the connect(); that will be changed later.)