Configure with --enable-dnstap by default
All platforms but OpenBSD have DNSTAP dependencies readily in their
respective repositories, and DNSTAP thus can be tested there. Given that
majority of images have DNSTAP dependencies available, it seems fitting
to make it the default.

For the sake of testing in Linux environments where DNSTAP dependencies
are not present, --disable-dnstap was added to Ubuntu Xenial job.