Run a minimal set of system tests on Windows for merge requests
Regular system test on Windows is run only for schedule-, web-, and
tag-originated pipelines because the test suite on Windows is slow and
takes scarce resources. However, sometimes BIND is broken on Windows,
and we find out only the day after the code was merged. To identify a
potential breakage, a minimal subset of the system test suite should be
run (only) for merge requests.