Remove the maximum allocation limit from the isc_mempool
The only place where the limits on the maximum number of allocated items
from the pool was the dns_dispatch where we already have different
limits in place.

As an example the maximum number of buffers is guarded by:

    if (disp->mgr->buffers >= DNS_DISPATCH_MAXBUFFERS) {
            UNLOCK(&disp->mgr->buffer_lock);
            return (NULL);
    }

but then at the same time we were limiting the maximum number of items
we can get from the disp->bpool.

By removing the maximum allocation limit from the isc_mempool API, we
can simplify the logic in many places as the isc_mempool_get() would
never fail now and it would always return a chunk of memory.