Increase wait_for_log timeout in kasp shell test
When running with TSAN in CI, the test ocassionally fails on:

'exceeded time limit waiting for literal 'keymgr: purgekeys.kasp done'
in ns4/named.run'

The line is actually present in the logs, but it take slightly longer
than 3 seconds to appear. Increase the wait_for_log timeout of the kasp
test to 10 seconds to avoid such issues.